+++ common/src/main/java/club/iananderson/seasonhud/client/Time.java
@@ -1,4 +1,0 @@
-package club.iananderson.seasonhud.client;
-
-public class Time {
-}