+++ fabric/src/main/java/club/iananderson/seasonhud/mixin/ftbchunks/FTBChunksClientMixin.java
@@ -7,6 +7,7 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
@@ -16,13 +17,14 @@ import java.util.List;
 import static club.iananderson.seasonhud.Common.SEASON_STYLE;
 import static club.iananderson.seasonhud.config.Config.enableMinimapIntegration;
 import static club.iananderson.seasonhud.config.Config.enableMod;
-import static club.iananderson.seasonhud.impl.seasons.CurrentSeason.getSeasonName;
+import static club.iananderson.seasonhud.impl.seasons.CurrentSeason.getSeasonHudName;
 import static dev.ftb.mods.ftbchunks.client.FTBChunksClientConfig.MINIMAP;
 
 @Mixin(FTBChunksClient.class)
 public class FTBChunksClientMixin {
 
-    private static BooleanValue MINIMAP_SEASON = (BooleanValue)MINIMAP.addBoolean("season",true).comment(new String[]{"Show season under minimap"});
+    @Unique
+    private static BooleanValue MINIMAP_SEASON = MINIMAP.addBoolean("season",true).comment(new String[]{"Show season under minimap"});
 
     @Inject(
             method = "buildMinimapTextData",
@@ -32,8 +34,8 @@ public class FTBChunksClientMixin {
 
     private void buildMinimapTextData(Minecraft mc, double playerX, double playerY, double playerZ, MapDimension dim, CallbackInfoReturnable<List<Component>> cir) {
         MutableComponent seasonCombined = Component.translatable("desc.seasonhud.combined",
-                getSeasonName().get(0).copy().withStyle(SEASON_STYLE),
-                getSeasonName().get(1).copy());
+                getSeasonHudName().get(0).copy().withStyle(SEASON_STYLE),
+                getSeasonHudName().get(1).copy());
         List<Component>res = cir.getReturnValue();
 
         enableMod.set(MINIMAP_SEASON.get());