+++ forge/src/main/java/club/iananderson/seasonhud/client/minimaps/MapAtlases.java
@@ -1,6 +1,7 @@
 package club.iananderson.seasonhud.client.minimaps;
 
 import club.iananderson.seasonhud.config.Config;
+import club.iananderson.seasonhud.impl.minimaps.CurrentMinimap;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Font;
@@ -9,7 +10,6 @@ import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.world.effect.MobEffect;
-import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.client.gui.overlay.ForgeGui;
 import net.minecraftforge.client.gui.overlay.IGuiOverlay;
 import pepjebs.mapatlases.MapAtlasesMod;
@@ -18,12 +18,11 @@ import pepjebs.mapatlases.client.MapAtlasesClient;
 import pepjebs.mapatlases.config.MapAtlasesClientConfig;
 
 import static club.iananderson.seasonhud.Common.SEASON_STYLE;
-import static club.iananderson.seasonhud.impl.minimaps.CurrentMinimap.loadedMinimap;
-import static club.iananderson.seasonhud.impl.seasons.CurrentSeason.getSeasonName;
+import static club.iananderson.seasonhud.Common.mc;
+import static club.iananderson.seasonhud.impl.seasons.CurrentSeason.getSeasonHudName;
 
 public class MapAtlases implements IGuiOverlay{
     protected final int BG_SIZE = 64;
-    private final Minecraft mc = Minecraft.getInstance();
 
     private static void drawStringWithLighterShadow(GuiGraphics context, Font font, MutableComponent text, int x, int y) {
         context.drawString(font, text, x + 1, y + 1, 5855577, false);
@@ -45,10 +44,10 @@ public class MapAtlases implements IGuiOverlay{
     }
 
     public static void drawMapComponentSeason(GuiGraphics poseStack, Font font, int x, int y, int targetWidth, float textScaling) {
-        if (loadedMinimap("map_atlases")) {
+        if (CurrentMinimap.minimapLoaded("map_atlases")) {
             MutableComponent seasonCombined = Component.translatable("desc.seasonhud.combined",
-                    getSeasonName().get(0).copy().withStyle(SEASON_STYLE),
-                    getSeasonName().get(1).copy());
+                    getSeasonHudName().get(0).copy().withStyle(SEASON_STYLE),
+                    getSeasonHudName().get(1).copy());
 
             float globalScale = (float)(double)MapAtlasesClientConfig.miniMapScale.get();
             drawScaledComponent(poseStack, font, x, y, seasonCombined, textScaling / globalScale, targetWidth, (int)(targetWidth / globalScale));
@@ -56,7 +55,7 @@ public class MapAtlases implements IGuiOverlay{
     }
 
     public static boolean shouldDraw(Minecraft mc) {
-        if (loadedMinimap("map_atlases")) {
+        if (CurrentMinimap.minimapLoaded("map_atlases")) {
             if (mc.level == null || mc.player == null) {
                 return false;
             }
@@ -77,10 +76,10 @@ public class MapAtlases implements IGuiOverlay{
 
     @Override
     public void render(ForgeGui gui, GuiGraphics seasonStack, float partialTick, int screenWidth, int screenHeight) {
-        if(loadedMinimap("map_atlases") && shouldDraw(mc)) {
+        if(CurrentMinimap.minimapLoaded("map_atlases") && shouldDraw(mc)) {
             float textScaling = (float) (double) MapAtlasesClientConfig.minimapCoordsAndBiomeScale.get();
 
-            int textHeightOffset = 2;
+            float textHeightOffset = 2.0F;
             float globalScale = (float) (double) MapAtlasesClientConfig.miniMapScale.get();
             int actualBgSize = (int) (BG_SIZE * globalScale);
 
@@ -90,8 +89,9 @@ public class MapAtlases implements IGuiOverlay{
             seasonStack.pose().scale(globalScale, globalScale, 1);
 
             Anchoring anchorLocation = MapAtlasesClientConfig.miniMapAnchoring.get();
-            int x = anchorLocation.isLeft ? 0 : (int) (screenWidth / globalScale) - BG_SIZE;
-            int y = anchorLocation.isUp ? 0 : (int) (screenHeight / globalScale) - BG_SIZE;
+            int offset = 5;
+            int x = anchorLocation.isLeft ? offset : (int) (screenWidth / globalScale) - (BG_SIZE + offset);
+            int y = anchorLocation.isUp ? offset : (int) (screenHeight / globalScale) - (BG_SIZE + offset);
             x += (int) (MapAtlasesClientConfig.miniMapHorizontalOffset.get() / globalScale);
             y += (int) (MapAtlasesClientConfig.miniMapVerticalOffset.get() / globalScale);
 
@@ -117,18 +117,18 @@ public class MapAtlases implements IGuiOverlay{
 
             if (Config.enableMod.get()) {
                 if (MapAtlasesClientConfig.drawMinimapCoords.get()) {
-                    textHeightOffset += (10 * textScaling);
+                    textHeightOffset = (textHeightOffset + 10.0F * textScaling);
                 }
 
                 if (MapAtlasesClientConfig.drawMinimapChunkCoords.get()) {
-                    textHeightOffset += (10 * textScaling);
+                    textHeightOffset = (textHeightOffset + 10.0F * textScaling);
                 }
 
                 if (MapAtlasesClientConfig.drawMinimapBiome.get()) {
-                    textHeightOffset += (10 * textScaling);
+                    textHeightOffset = (textHeightOffset + 10.0F * textScaling);
                 }
 
-                drawMapComponentSeason(seasonStack, font, (int) (x), (int) (y + BG_SIZE + (textHeightOffset / globalScale)), actualBgSize, textScaling);
+                drawMapComponentSeason(seasonStack, font, x, (int)(y + BG_SIZE + (textHeightOffset / globalScale)), actualBgSize, textScaling);
                 seasonStack.pose().popPose();
             }
         }