+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
+## [1.8.0] - 2024-05-24
+
+### Changed
+- Updated dependencies to the latest versions
+- Rewrote a bit to allow for more common code between Forge and Fabric
+
+### Fixed
+-Fixed position with latest version of MapAtlases
+
+
 ## [1.7.19] - 2024-04-08
 
 ### Fixed